requirements.txt
fastapi==0.128.1
protobuf==6.33.5
pymongo==4.16.0
python-dotenv==1.2.1
Requests==2.32.5
